// slider2/types.ts
export interface SliderRangeOptions {
  /**
   * The min value for the slider.
   */
  min?: number;

  /**
   * The max value for the slider.
   */
  max?: number;

  /**
   * Boolean if the slider is rendered vertically instead of horizontally.
   */
  vertical?: boolean;

  /**
   * Boolean if the slider is disabled and the values cannot be changed.
   */
  disabled?: boolean;
}

export interface SliderStepOptions extends SliderRangeOptions {
  /**
   * A positive number representing the value to "jump" while incrementing or
   * decrementing the slider's value. This should normally stay as the default
   * value of `1`, but can also be decimal values if needed.
   */
  step?: number;
}

export interface SliderStateOptions extends SliderRangeOptions {
  /**
   * Boolean if the dense spec has been applied.
   */
  dense?: boolean;

  /**
   * A function that is used to help with accessibility by creating a better
   * value string if just a number isn't representative enough of your range.
   */
  getValueText?(value: number): string;

  /**
   * Boolean if the slider is currently being dragged. This is mostly used to
   * disable the easing animation while being dragged.
   */
  dragging?: boolean;

  /**
   * Boolean if the transition should be disabled. This should not really be
   * touched since this is normally handled by the `useSlider` hook.
   */
  disableTransition?: boolean;
}

export type SliderEventHandlerNames =
  | "onKeyUp"
  | "onKeyDown"
  | "onMouseDown"
  | "onTouchStart";

export type SliderEventHandlers = Pick<
  HTMLAttributes<HTMLSpanElement>,
  SliderEventHandlerNames
>;